requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
dataclass-wizard==0.35.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
frozenlist==1.7.0
fsspec==2024.12.0
greenlet==3.2.4
grpcio==1.74.0
h11==0.16.0
h2==4.2.0
h5py==3.14.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
json_repair==0.44.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-mongodb==0.6.2
langchain-neo4j==0.5.0
langchain-openai==0.3.28
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.9
langgraph==0.6.5
langgraph-checkpoint==2.1.1
langgraph-checkpoint-mongodb==0.1.4
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.0
langsmith==0.4.13
lark==1.2.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mem0ai==0.1.116
motor==3.7.1
multidict==6.6.3
mypy_extensions==1.1.0
neo4j==5.28.2
neo4j-graphrag==1.9.1
numpy==2.3.2
numpy-quaternion==2024.0.10
openai==1.98.0
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
pillow==11.3.0
portalocker==3.2.0
posthog==6.6.1
propcache==0.3.2
protobuf==5.29.5
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.12.1
pypdf==5.9.0
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.15.1
rank-bm25==0.2.2
redis==6.4.0
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rq==2.4.1
scipy==1.16.1
sentry-sdk==2.34.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.10.0
tqdm==4.67.1
typer==0.16.0
types-PyYAML==6.0.12.20250822
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
